pyproject.toml
[tool.poetry]
authors = ["Multisafepay <integration@multisafepay.com>"]
description = "Multisafepay API wrapper"
license = "OSL-3.0"
name = "multisafepay"
version = "0.1.0"

[tool.poetry.dependencies]
python = ">=3.8,<3.14"
requests = "^2.32.3"
toml = "^0.10.2"
pydantic = "^1.10.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
mypy = "^1.10.0"
pylint = "^3.2.2"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
ruff = "^0.4.4"
pytest-rerunfailures = "^12.0"

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

# configures some basic black settings per this article:
# https://ljvmiranda921.github.io/notebook/2018/06/21/precommits-using-black-and-flake8/
[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.env
  | _build
  | buck-out
  | build
  | dist
  | env
)/
'''
include = '\.pyi?$'
line-length = 79

[tool.mypy]
disable_error_code = [
    "import-untyped",
    "assignment",
    "return-value",
    "return",
    "union-attr",
    "index",
    "override",
    "arg-type",
]

[tool.pylint."MESSAGE CONTROL"]
disable = [
    "C0114",
    "C0301",
    "C0411",
    "C0415",
    "C2801",
    "E0213",
    "E0611",
    "E1101",
    "E1133",
    "E1136",
    "R0801", # Disables duplicate code warning
    "R0902",
    "R0903", # Disables too few public methods
    "R0904",
    "R0913",
    "R0917",
    "W0102",
    "W0221",
    "W0246",
    "W0612",
    "W0613",
    "W0622",
    "W0707",
    "W0718",
    "W1514",
]

[tool.ruff.lint]
extend-safe-fixes = [
    "D400", # docstrings should end with a period
    "D415", # docstrings should end with a period, question mark, or exclamation point
]
ignore = [
    "A002",
    "ANN001",
    "ANN003",
    "ANN101", # missing type annotation for self
    "ANN102", # missing type annotation for cls
    "ANN201",
    "ANN204",
    "ANN401",
    "ARG002",
    "B006",
    "B904",
    "BLE001",
    "D100",
    "D101",
    "D103",
    "D104",
    "D107",
    "D200",
    "D203", # no blank line before class
    "D205",
    "D212", # multi-line summary first line
    "D401",
    "D417",
    "DTZ006",
    "DTZ007",
    "E501",
    "E712",
    "EM101",
    "EM102",
    "F811",
    "F841",
    "FA", # flake8-future-annotations ruleset
    "FBT001",
    "FBT002",
    "FBT003",
    "INP001",
    "N805",
    "N818",
    "PLR0913",
    "PLR2004",
    "PT011",
    "PT017",
    "PTH123",
    "RUF012",
    "RUF013",
    "S105",
    "S106",
    "T201",
    "TCH002",
    "TID252",
    "TRY003",
    "TRY300",
]
select = ["ALL"]
